Move read_default_host_dir() to its own stdlib-only module
config/host_dir.py has no third-party imports, so all callers --
including the lightweight tab-completion entrypoint (complete.py) --
can import it directly. This eliminates the last inline copy of the
host-dir resolution logic.
